Add files via upload
New logo for the app tray as previous app utilized an icon built into Google. This made the icon very pixelated and didn't look nice.

Also changed the input on each of the user input fields to numberSigned|numberDecimal as before it was only numberDecimal, meaning that users could not implement negative numbers for use with the provided basic operators.